reorganize new project wizard directory structure
