+++ src/vs/workbench/browser/positronNewProjectWizard/interfaces/newProjectWizardStepProps.ts
@@ -2,7 +2,7 @@
  *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
 
-import { NewProjectWizardStep } from 'vs/workbench/browser/positronModalDialogs/newProjectWizard/steps/newProjectWizardStep';
+import { NewProjectWizardStep } from 'vs/workbench/browser/positronNewProjectWizard/interfaces/newProjectWizardStep';
 
 /**
  * The NewProjectWizardStepProps interface provides the wizard navigation functions
