+++ src/vs/workbench/browser/positronNewProjectWizard/steps/projectTypeStep.css
